Modify new hash to properly utilize crypt's features
The naive swap I made earlier ignored the fact that the crypt return
value includes the key factor and salt values which allows a stored
hash to readily be compared using the crypt method. This is the
preferred option as it allows either of these values to be changed
without invalidating current passwords. (For instance, increasing the
key factor in a couple years to keep up with performance improvements
of typical computers.)