Create integer class
Mostly this class exists to have a canonical place for the
positiveIntegerForValue method which is a common operation I do when
setting numeric id properties within an object. This class would
actually be useful if PHP had a way for objects to return their value
for anything other than a string context.