Swap the context of keys in the changed list
I had originally used the local key for changed values, however this
ignores the fact that one aspect of the change is that the value may
have also moved. Thinking of the difference collection as something to
be applied to the local variable leads me to beleive the key should
correspond to the comparison object instead.